+++ src/field_5x52_int128_impl.h
@@ -42,7 +42,7 @@ void static inline secp256k1_fe_mul_inner(const uint64_t *a, const uint64_t *b, 
     uint64_t t7 = c & 0xFFFFFFFFFFFFFULL; c = c >> 52; // c max 02000000000000FE
     c = c + (__int128)a[4] * b[4];
     uint64_t t8 = c & 0xFFFFFFFFFFFFFULL; c = c >> 52; // c max 001000000000001E
-    uint64_t t9 = c;
+    uint64_t t9 = (uint64_t)c;
 
     c = t0 + (__int128)t5 * 0x1000003D10ULL;
     t0 = c & 0xFFFFFFFFFFFFFULL; c = c >> 52; // c max 0000001000003D10
@@ -56,7 +56,7 @@ void static inline secp256k1_fe_mul_inner(const uint64_t *a, const uint64_t *b, 
     r[4] = c & 0x0FFFFFFFFFFFFULL; c = c >> 48; // c max 000001000003D110
     c = t0 + (__int128)c * 0x1000003D1ULL;
     r[0] = c & 0xFFFFFFFFFFFFFULL; c = c >> 52; // c max 1000008
-    r[1] = t1 + c;
+    r[1] = t1 + (uint64_t)c;
 
 }
 
@@ -85,7 +85,7 @@ void static inline secp256k1_fe_sqr_inner(const uint64_t *a, uint64_t *r) {
     uint64_t t7 = c & 0xFFFFFFFFFFFFFULL; c = c >> 52; // c max 02000000000000FE
     c = c + (__int128)a[4] * a[4];
     uint64_t t8 = c & 0xFFFFFFFFFFFFFULL; c = c >> 52; // c max 001000000000001E
-    uint64_t t9 = c;
+    uint64_t t9 = (uint64_t)c;
     c = t0 + (__int128)t5 * 0x1000003D10ULL;
     t0 = c & 0xFFFFFFFFFFFFFULL; c = c >> 52; // c max 0000001000003D10
     c = c + t1 + (__int128)t6 * 0x1000003D10ULL;
@@ -98,7 +98,7 @@ void static inline secp256k1_fe_sqr_inner(const uint64_t *a, uint64_t *r) {
     r[4] = c & 0x0FFFFFFFFFFFFULL; c = c >> 48; // c max 000001000003D110
     c = t0 + (__int128)c * 0x1000003D1ULL;
     r[0] = c & 0xFFFFFFFFFFFFFULL; c = c >> 52; // c max 1000008
-    r[1] = t1 + c;
+    r[1] = t1 + (uint64_t)c;
 
 }
